// src/shared/types/mock-server-data.type.ts
export type MockServerData = {
  name: string[];
  description: string[];
  city: string[];
  previewImage: string[];
  photos: string[];
  isPremium: string[];
  isFavorite: string[];
  housingType: string[];
  amenities: string[];
  author: string[];
  coordinates: {
      latitude: number;
      longitude: number;
  }[];
}